pyproject.toml
[project]
name = "mcp-nta"
version = "0.1.0"
description = "MCP server for Ireland's NTA public transport — focused, queryable tools with human-readable output. Inspired by ireland-nta-mcp."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "gtfs-realtime-bindings>=1.0.0",
    "mcp[cli]>=1.12.4",
    "protobuf>=6.31.1",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pyright>=1.1.389",
    "ruff>=0.7.3",
]

[project.scripts]
mcp-nta = "mcp_nta:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["mcp_nta"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
